requirements.txt
openpyxl==3.0.0
pandas==1.3.2
pandas-datareader==0.9.0
pyarrow==4.0.0
PyYAML==5.4.1
requests==2.22.0
wget==3.2
matplotlib==3.5.2
statsmodels==0.13.2
numpy==1.22.4
scipy==1.8.0
statsmodels==0.13.2
graphviz==0.20
numba==0.55.2
sequence-jacobian==1.0.0
